% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{get_swCloud_daily}
\alias{get_swCloud_daily}
\title{get_swCloud_daily}
\usage{
get_swCloud_daily(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud_daily}} or \code{\linkS4class{swInputData}}.}
}
\description{
get_swCloud_daily
}
\seealso{
\code{\linkS4class{swCloud_daily}} and \code{\linkS4class{swInputData}}
}
