% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swCloud_daily<-}
\alias{set_swCloud_daily<-}
\title{set_swCloud_daily<-}
\usage{
set_swCloud_daily(object) <- value
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud_daily}} or \code{\linkS4class{swInputData}}.}

\item{value}{A value to assign to a specific slot of the \code{object}.}
}
\description{
set_swCloud_daily<-
}
\seealso{
\code{\linkS4class{swCloud_daily}} and \code{\linkS4class{swInputData}}
}
