% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{set_swCloud_daily}
\alias{set_swCloud_daily}
\title{set_swCloud
Need to define and export this generic method -- otherwise, \code{\link{set_swCloud_daily<-}} doesn't work.}
\usage{
set_swCloud_daily(object, value)
}
\description{
set_swCloud
Need to define and export this generic method -- otherwise, \code{\link{set_swCloud_daily<-}} doesn't work.
}
