% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D_swCloud_daily.R
\docType{class}
\name{swCloud_daily-class}
\alias{swCloud_daily-class}
\alias{initialize,swCloud_daily-method}
\alias{get_swCloud_daily,swCloud_daily-method}
\alias{swCloud_daily_Humidity,swCloud_daily-method}
\alias{set_swCloud_daily<-,swCloud_daily-method}
\alias{swCloud_daily_Humidity<-,swCloud_daily-method}
\alias{swReadLines,swCloud_daily,character-method}
\title{Class \code{"swCloud_daily"}}
\usage{
\S4method{initialize}{swCloud_daily}(.Object, ...)

\S4method{get_swCloud_daily}{swCloud_daily}(object)

\S4method{swCloud_daily_Humidity}{swCloud_daily}(object)

\S4method{set_swCloud_daily}{swCloud_daily}(object) <- value

\S4method{swCloud_daily_Humidity}{swCloud_daily}(object) <- value

\S4method{swReadLines}{swCloud_daily,character}(object, file)
}
\arguments{
\item{.Object}{An object of class \code{\linkS4class{swCloud_daily}}.}

\item{...}{Further arguments to methods.}

\item{object}{An object of class \code{\linkS4class{swCloud_daily}}.}

\item{value}{A value to assign to a specific slot of the object.}

\item{file}{A character string. The file name from which to read.}
}
\description{
The methods listed below work on this class and the proper slot of the class
  \code{\linkS4class{swInputData}}.
}
\examples{
showClass("swCloud_daily")
x <- new("swCloud_daily")

}
\seealso{
\code{\linkS4class{swInputData}} \code{\linkS4class{swFiles}}
\code{\linkS4class{swWeather}} \code{\linkS4class{swInputData}}
\code{\linkS4class{swMarkov}} \code{\linkS4class{swProd}}
\code{\linkS4class{swSite}} \code{\linkS4class{swSoils}}
\code{\linkS4class{swEstab}} \code{\linkS4class{swOUT}}
\code{\linkS4class{swSWC}} \code{\linkS4class{swLog}}
}
