% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_swGenericMethods.R
\name{swCloud_daily_Humidity}
\alias{swCloud_daily_Humidity}
\title{swCloud_Humidity}
\usage{
swCloud_daily_Humidity(object)
}
\arguments{
\item{object}{An object of class \code{\linkS4class{swCloud_daily}} or \code{\linkS4class{swInputData}}.}
}
\description{
swCloud_Humidity
}
\seealso{
\code{\linkS4class{swCloud_daily}} and \code{\linkS4class{swInputData}}
}
